Separate some logic to different components
