+++ client/src/utils.js
@@ -3,6 +3,13 @@ import { server_url } from './config.json'
 
 export const log_event = new CustomEvent('user-log');
 
+/**
+ * request() make an AJAX request to server and return promise with result
+ * Parameters:
+ * - path {string} - route to which the request is made (e.g. '/event')
+ * - method {string} - request method, GET by default
+ * - data {json} - json with data to be sent
+ */
 export async function request(path="", method="GET", data=null) {
     let myHeaders = new Headers();
     myHeaders.append('Content-Type', 'application/json');
